[coln-query] cleanups
